+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* **feature** Added isOptedOut parameter to the PiwikTracker. [#124](https://github.com/piwik/piwik-sdk-ios/issues/124)
 
 ## 4.0.0-beta2
 * **feature** Added the possibility to set the url for a sreen view event. [#92](https://github.com/piwik/piwik-sdk-ios/issues/92)