+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* **feature** Renamed the Tracker to PiwikTracker. [#146](https://github.com/piwik/piwik-sdk-ios/issues/146)
 
 ## 4.0.0-beta2
 * **feature** Added the possibility to set the url for a sreen view event. [#92](https://github.com/piwik/piwik-sdk-ios/issues/92)